requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
beartype==0.14.1
better-abc==0.0.3
certifi==2026.1.4
charset-normalizer==3.4.4
circuitsvis==1.43.3
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.5.0
debugpy==1.8.20
decorator==5.2.1
dill==0.4.0
einops==0.8.2
executing==2.2.1
fancy-einsum==0.0.3
fastjsonschema==2.21.2
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jaxtyping==0.3.7
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.16.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==1.26.4
packaging==26.0
pandas==3.0.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.5.1
plotly==6.5.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.17.0
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.52.0
setuptools==82.0.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformer-lens==2.17.0
transformers==4.57.6
transformers-stream-generator==0.0.5
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wadler_lindig==0.1.7
wandb==0.24.2
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
